fix: widen oracle_update_cadence columns to NUMERIC(78,0) for Chainlink uint80
BIGINT max is 9.2e18; Chainlink round_id is uint80 (max 1.2e24).
Migration 095 ALTERs both round_id and updated_at_block to NUMERIC(78,0).
Startup DDL and migration 093 source updated to match.

Collector already uses Python int (arbitrary precision) — no code change
needed. psycopg2 passes Python int to NUMERIC cleanly.

+++ migrations/093_oracle_update_cadence.sql
@@ -4,9 +4,9 @@
 CREATE TABLE IF NOT EXISTS oracle_update_cadence (
     id BIGSERIAL PRIMARY KEY,
     oracle_id TEXT NOT NULL,
-    round_id BIGINT NOT NULL,
+    round_id NUMERIC(78,0) NOT NULL,
     answer NUMERIC,
-    updated_at_block BIGINT NOT NULL,
+    updated_at_block NUMERIC(78,0) NOT NULL,
     updated_at_timestamp TIMESTAMPTZ NOT NULL,
     observed_at TIMESTAMPTZ NOT NULL DEFAULT NOW(),
     gap_from_previous_seconds INTEGER,
